Risc-I added support to load and save instruction etc.
The circuit now supports all save and loading instructions, excluding the relative ones.
The control instructions are also working, but CALLI and the relative are not implemented yet
The ula flags output has been fixed, as it was reversed before, as well as the memory connection that was also reversed. some other circuit bugs were fixed too.