DRISC Assembler
Added a new assembler algorithm that compiles code in the RISC-V format, with the same instructions and register name parsing. it also prints the assembly steps in a simpler and color coded way to make debbugging easier.

The old assembler is still here and to activate it you need to activate in the config file changing the version from DRisc to anything else