Pipeline changed and other improvements made
the pipeline was modified so that load/store instructions third phase runs in parallel with the first and second phases of the next instructions
for example, now the pipeline runs like this:
D=Decode X=Execute S= Store or load
   D-X-S
       D-X
Although this makes the architecture more efficient it also introduced a hazard for Load instructions, where the following instruction will only have access to the old value of the destination register of the previous Load instruction. This hazard was introduced as a warning in the previous assembler commit and should be considered for future programs until a better solution has been implemented.

Branch and Jump instructions now have a pipeline flush right after their execution to prevent hazards. Although this is not the most efficient solution, there is no compiler to help with those hazards automatically, so its better to consider a solutions that makes it easier for the user to create a program. For future updates it would be nice to consider branch prediction to avoid the need for the pipeline flush.

Other minor updates are
-Better, more user friendly visuals for the components
-Simplified the functionality of the operation controllers by removing the result registers and the input to enable the result output. This way the components always outputs values that will need to be deactivated or multiplexed externally.
-Separation of the DataIO and the shifter.  For pratical reasons this may not be the best idea, but now the Data IO doesn't use the shifter on the load instructions, instead it has its own circuit that does that.  It may use more components and gates, but it simplifies the shifter circuit(which may be fused with the ALU in future updates) and also looks simpler in the simulation.
-Changed the first register of the program counter into a counter, removing the necessity of other components to increment this register.
-lots of other minor changes were done to the controller and the connection to the external IO