Circuit conditions and timing update
Alu overflow detection has been changed for both RISC-I and DRISC for a more precise model.

the Shifter read signal was simplified because the data IO does not use the shifter anymore. the OP_function may also be changed later for the same reason.

All the clocks have been inverted so that most of the circuit updates at positive edge. This change makes everything be more in line with what is usually used and also makes it more consistent with what was defined in the SystemVerilog code.

The phases went back to being generated by a ring counter. this removes the necessity for external phase input.

finally although it may not make a difference, the write in the next pc counter is now constrained by the jump, the same way as in the verilog definition