Added support to macros and multinstruction pseudoinstructions
now the assembler support macros with the following syntax:

.macro macroName argument1 argument2 ....
    macro commands
.endmacro

The assembler also supports pseudoinstructions that are translated to more than one instruction.  Two examples of that were the li(load immediate) and Call instructions. the LI instruction is a new one that should simplify loading constants in the assembly.

The supported extension changed from .asm to .dasm to differenciate it from regular asm. the resulting code extension is .mem instead of .asm.