new system IO
The system IO has been updated to a more patronized system. There is a new component that based on the current address and a range, controls the read and write signal that is passed through it. This allows the an easier selection of an IO device based on an address. To demonstrate that some IO devices have been configured. There is a screen were each pixel is mapped from address 0x1000000 to 0x1100000 that can be written to, and there is a keyboard, joystick, switches and random number generator mapped between the addresses 0xfffffd and 0xffffff.

The capabilities of the keyboard and the screen are demonstrated by the UserIO program. This program also makes use of macros and the new li instruction