SV Simulation integration with Interaction program
A new program was created to communicate with the System Verilog simulation.  It works by writing the keyboard input into a file that is read by the simulation. The simulation can also write into an address that controls an array of pixels that is read by the new program. This allows the simulation to be interactive

for this integration two components were created to simulate the user input and screen controllers.