Operation Controller, modified registers update timing
now current and last instruction registers update on the rising edge of the clock and on the phase 3. This was changed for some reasons. The first was to keep consistent with other registers. The second reason has to do with propagation delays and how the instructions were being displayed and computed; before, the cycle and current instruction did not match, even though the current instruction was suposed to be updated at the first phase, and this caused some hard to understand behavior. This was because the phase only updates with a small delay after the rising edge of the clock, so the rising edge of the clock matches the previous phase, not the next one, so if we want the current instruction to match the rising clock signal, we need to use the previous phase, which in our case is phase 3.  The diagram below tries summarizes the timing changes

before
clock         __/¨¨\__/¨¨\__/¨¨\__/¨¨\__/¨¨\__/¨¨\
phase        ---1-----2----3-----1-----2-----3--
next Inst   ----------2-----------------3-------
curr Inst    ------1-----------------2-----------

now:

before
clock         __/¨¨\__/¨¨\__/¨¨\__/¨¨\__/¨¨\__/¨¨\
phase        ---1-----2----3-----1-----2-----3--
next Inst   ----------2-----------------3-------
curr Inst    ---1------------------2-------------