DRISC Pipeline Optimization
The pipeline clock cycles count per phase was reduced from 3 to 2, improving the speed of every instruction.

This came with the removal of some registers in favor of connecting the information source directly into the target. Those are the pipeline flush flip flop, the Address register and the data io output register.

The B bus now connects directly with the data bus, and whenever there is a write signal, the data from B bus is sent to the data bus.

These changes simplified the data IO, the phase generator and some of the control signals

data IO has been renamed to Input Buffer since it does not affects output anymore